refactor: Consolidate utility functions into dedicated packages (i906)
This commit refactors the project structure by moving various utility
functions into more appropriate packages. This improves code organization
and maintainability.

Specifically:
- Icon pack related utility functions (`updateIconPackInfos`,
  `cacheIconPackFile`) have been moved from `domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/iconpack`
  to `domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/util`.
- Grid item utility functions (`cleanupGridItemRecursively`,
  `getFolderGridItemsById`, `updatePinShortcutsByPackageName`) have been
  moved from various locations to
  `domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/util/GridItemUtil.kt`.
- References to these moved functions have been updated accordingly.

This change centralizes related utilities, making them easier to find
and manage.

Affected files:
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/launcherapps/SyncDataUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/launcherapps/ChangePackageUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/page/UpdatePageItemsUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/DeleteGridItemUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/UpdateGridItemsAfterMoveUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/launcherapps/LauncherAppsUtil.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/launcherapps/AddPackageUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/iconpack/UpdateIconPackInfosUseCase.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/iconpack/IconPackInfoUseCaseUtil.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/grid/GridUtil.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/util/IconPackInfoUtil.kt
- domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/util/GridItemUtil.kt
