+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/page/UpdatePageItemsUseCase.kt
@@ -28,7 +28,7 @@ import com.eblan.launcher.domain.model.PageItem
 import com.eblan.launcher.domain.repository.FolderGridItemRepository
 import com.eblan.launcher.domain.repository.GridRepository
 import com.eblan.launcher.domain.repository.UserDataRepository
-import com.eblan.launcher.domain.usecase.grid.cleanupGridItemRecursively
+import com.eblan.launcher.domain.usecase.util.cleanupGridItemRecursively
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.withContext
