+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/util/IconPackInfoUtil.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.domain.usecase.iconpack
+package com.eblan.launcher.domain.usecase.util
 
 import com.eblan.launcher.domain.common.IconKeyGenerator
 import com.eblan.launcher.domain.framework.FileManager
